Simplify QR code component
