+++ src/library/catalog.ts
@@ -33,17 +33,17 @@ export interface LabComponentDefinition {
 
 export const componentCatalog: LabComponentDefinition[] = [
   {
-    id: 'qr-code-card',
+    id: 'qr-code',
     name: {
-      en: 'QR Code Card',
-      ko: 'QR 코드 카드',
+      en: 'QR Code',
+      ko: 'QR 코드',
     },
     description: {
-      en: 'Generate a branded QR card with custom colors and a centered logo.',
-      ko: '사용자 지정 색상과 중앙 로고를 포함한 QR 카드입니다.',
+      en: 'A clean QR code with optional center icon and customizable colors.',
+      ko: '중앙 아이콘과 색상을 자유롭게 조절할 수 있는 단순한 QR 코드입니다.',
     },
-    component: () => import('@/components/library/QrCodeCard.vue'),
-    preview: () => import('@/demos/QrCodeCardDemo.vue'),
+    component: () => import('@/components/library/QrCode.vue'),
+    preview: () => import('@/demos/QrCodeDemo.vue'),
     defaultProps: {
       content: 'https://component-lab.dev/welcome',
       lightColor: '#ffffff',
